Move aggregator API into sdk/export/metric/aggregator
