+++ exporter/metric/stdout/stdout.go
@@ -24,7 +24,7 @@ import (
 	"time"
 
 	export "go.opentelemetry.io/otel/sdk/export/metric"
-	"go.opentelemetry.io/otel/sdk/metric/aggregator"
+	"go.opentelemetry.io/otel/sdk/export/metric/aggregator"
 )
 
 type Exporter struct {
