--- sdk/metric/aggregator/api.go
@@ -1,54 +1,0 @@
-// Copyright 2019, OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package aggregator // import "go.opentelemetry.io/otel/sdk/metric/aggregator"
-
-import (
-	"time"
-
-	"go.opentelemetry.io/otel/api/core"
-)
-
-type (
-	Sum interface {
-		Sum() (core.Number, error)
-	}
-
-	Count interface {
-		Count() (int64, error)
-	}
-
-	Max interface {
-		Max() (core.Number, error)
-	}
-
-	Quantile interface {
-		Quantile(float64) (core.Number, error)
-	}
-
-	LastValue interface {
-		LastValue() (core.Number, time.Time, error)
-	}
-
-	MaxSumCount interface {
-		Sum
-		Count
-		Max
-	}
-
-	Distribution interface {
-		MaxSumCount
-		Quantile
-	}
-)
