+++ sdk/metric/aggregator/ddsketch/ddsketch.go
@@ -23,7 +23,7 @@ import (
 	"go.opentelemetry.io/otel/api/core"
 
 	export "go.opentelemetry.io/otel/sdk/export/metric"
-	"go.opentelemetry.io/otel/sdk/metric/aggregator"
+	"go.opentelemetry.io/otel/sdk/export/metric/aggregator"
 )
 
 // Config is an alias for the underlying DDSketch config object.
